requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic==2.10.6
python-dotenv==1.0.1
google-cloud-aiplatform==1.81.0
httpx==0.28.1
